SOUND: Move FEVFile class to sound
