+++ src/aurora/rules.mk
@@ -69,7 +69,6 @@ src_aurora_libaurora_la_SOURCES += \
     src/aurora/thewitchersavefile.h \
     src/aurora/thewitchersavewriter.h \
     src/aurora/sacfile.h \
-    src/aurora/fevfile.h \
     src/aurora/gfxfile.h \
     $(EMPTY)
 
@@ -116,7 +115,6 @@ src_aurora_libaurora_la_SOURCES += \
     src/aurora/thewitchersavefile.cpp \
     src/aurora/thewitchersavewriter.cpp \
     src/aurora/sacfile.cpp \
-    src/aurora/fevfile.cpp \
     src/aurora/gfxfile.cpp \
     $(EMPTY)
