+++ src/sound/rules.mk
@@ -35,6 +35,7 @@ src_sound_libsound_la_SOURCES += \
     src/sound/xactsoundbank_binary.h \
     src/sound/fmodsamplebank.h \
     src/sound/wwisesoundbank.h \
+    src/sound/fevfile.h \
     $(EMPTY)
 
 src_sound_libsound_la_SOURCES += \
@@ -49,6 +50,7 @@ src_sound_libsound_la_SOURCES += \
     src/sound/xactsoundbank_binary.cpp \
     src/sound/fmodsamplebank.cpp \
     src/sound/wwisesoundbank.cpp \
+    src/sound/fevfile.cpp \
     $(EMPTY)
 
 src_sound_libsound_la_LIBADD = \
